% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{compare_last_names}
\alias{compare_last_names}
\title{Vectorised compound-name comparator}
\usage{
compare_last_names(names1, names2)
}
\arguments{
\item{names1, names2}{Character vectors of equal length.}
}
\value{
A numeric vector of similarities in \verb{[0, 1]}.
}
\description{
The comparator object passed to \code{\link[reclin2:compare_pairs]{reclin2::compare_pairs()}} for the surname
field. Applies \code{\link[=compare_two_names]{compare_two_names()}} pairwise over two equal-length character
vectors.
}
\examples{
compare_last_names(c("MCLANE", "SMITH"), c("MC-LANE", "SMYTHE"))
}
