% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{create_emp_ids}
\alias{create_emp_ids}
\title{Hash a vector of strings into deterministic EMP identifiers}
\usage{
create_emp_ids(v)
}
\arguments{
\item{v}{A character vector.}
}
\value{
A character vector of \code{EMP-...} identifiers, one per element of \code{v}.
}
\description{
Hash a vector of strings into deterministic EMP identifiers
}
\examples{
create_emp_ids(c("a", "b", "a"))
}
