% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_features}
\alias{model_features}
\title{The feature (comparison) columns of a comparisons data frame}
\usage{
model_features(comparisons)
}
\description{
The feature (comparison) columns of a comparisons data frame
}
\keyword{internal}
